Move function into render.py.
